requirements.txt
wheel
pandas==1.3.4
PyYAML==6.0
Scrapy==2.5.1
scrapy-splash==0.8.0
